pyproject.toml
# pyproject.toml

[tool.poetry]
name = "multi_tenant_application"
version = "0.1.0"
description = "A Django multi-tenant project"
authors = ["Oran Ben-Zur"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.2.8"
djangorestframework = "^3.12.4"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.1"
django-dynamic-db-router = "^0.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
mixer = "^7.2.2"
pytest-django = "^4.4.0"
